requirements.txt
apify_client==2.3.0
discord-webhook==1.4.1
openai==2.15.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
typing-inspection==0.4.2
typing_extensions==4.15.0
